fix refresh code
